Separate out ACL and backup schema into another file that is only included for non-oss builds
